--- frontend/src/components/Metrics.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function Metrics() {
-  return <h1>METRICS PAGE</h1>;
-}
