+++ frontend/src/index.js
@@ -1,6 +1,6 @@
 import { createRoot } from "react-dom/client";
 
-import App from "./components/App";
+import App from "./components/App/App";
 
 // cookies
 import { CookiesProvider } from "react-cookie";
